Fixed issues found by PVS-Studio (#729)
* Correctly compute SoftBodyShape memory size in bytes
* Remove empty loop with no side effects
* Drop already tru condition guaranteed above
* Ensure no out of bounds read/write to triangles in production JPH_ASSERT is in Debug only. Release builds can use UINT_MAX as result.  It leads to out of bounds reads and sometimes not crash but hides bugs. Suggest to hard stop in this case instead of failing silently.
* At least check in Debug mode we do not overflow during first vertex computation
* Correctly print size_t. Unfortunately MinGW doesn't support %zu C11 specifier. So decided to cast size_t -> uint.
* Correct format specifiers for printf-like API
* Create objects in place to speedup containers filling
* Speedup reading pointer data

+++ Jolt/Geometry/EPAPenetrationDepth.h
@@ -201,7 +201,7 @@ public:
 		hull.DrawLabel("Build initial hull");
 #endif
 #ifdef JPH_EPA_PENETRATION_DEPTH_DEBUG
-		Trace("Init: num_points = %d", support_points.mY.size());
+		Trace("Init: num_points = %u", (uint)support_points.mY.size());
 #endif
 		hull.Initialize(0, 1, 2);
 		for (typename Points::size_type i = 3; i < support_points.mY.size(); ++i)
